test(e2e): harden MCP compatibility review coverage
